fix: refactor original function for cost booking
