+++ src/customerUtils.js
@@ -3,6 +3,22 @@ const displayPrevBookedRooms = (customerObject, bookingsData) => {
   return bookingsData.filter((book) => book.userID === customerObject.id)
 }
 
+const calculateBookingCost = (displayPrevBookedRoomsData, roomDataInformation) => {
+  let newArray = []
+  let sum = 0;
+  roomDataInformation.filter((room) => {
+    displayPrevBookedRoomsData.filter((booking) => {
+      if (booking.roomNumber === room.number) {
+        newArray.push(room)
+      }
+    })
+  })
+  newArray.forEach((fil) => {
+    sum += fil.costPerNight
+  })
+  return sum
+}
+
 const findValidIDNumber = (customerData, username) => {
   let customerLogin;
   customerData.forEach((customer) => {
@@ -15,5 +31,6 @@ const findValidIDNumber = (customerData, username) => {
 
 export {
   displayPrevBookedRooms,
-  findValidIDNumber
+  findValidIDNumber,
+  calculateBookingCost
 }
