+++ src/domUpdates.js
@@ -1,6 +1,6 @@
-import { displayPrevBookedRooms, findValidIDNumber } from '../src/customerUtils.js'
+import { displayPrevBookedRooms, findValidIDNumber, calculateBookingCost} from '../src/customerUtils.js'
 import { findAvailability } from '../src/bookingUtils.js'
-import { filterByRoomType, calculateBookingCost } from '../src/roomsUtils.js'
+import { filterByRoomType } from '../src/roomsUtils.js'
 import { getAllCustomers, getSingleCustomer, getAllRooms, getAllBookings, addPostBooking } from '/Users/hollisvohr/turing_work/mod_2/outlook-project/src/apiCalls.js'
 
 //Query Selectors
