+++ src/roomsUtils.js
@@ -4,18 +4,19 @@ const filterByRoomType = (roomsData, roomType) => {
   return find
 }
 
-const calculateBookingCost = (roomsData, roomCosts) => {
-  let sum = 0;
-  roomsData.forEach((room) => {
-    if (roomCosts.includes(room.costPerNight)) {
-      sum += room.costPerNight
-    }
-  })
-  return sum
-}
+// const calculateBookingCost = (bookingData, roomsData) => {
+//   let sum = 0;
+//   let find = bookingData.filter((book) => book.userID === customerObject.id)
+//   roomsData.forEach((room) => {
+//     if (roomCosts.includes(room.costPerNight)) {
+//       sum += room.costPerNight
+//     }
+//   })
+//   return sum
+// }
 
 
 export {
   filterByRoomType,
-  calculateBookingCost
+  // calculateBookingCost
 }
