+++ test/roomsUtils-test.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 const expect = chai.expect;
 import { roomsData } from '/Users/hollisvohr/turing_work/mod_2/outlook-project/src/data/rooms-sample-data.js';
-import { filterByRoomType, calculateBookingCost } from '../src/roomsUtils.js'
+import { filterByRoomType } from '../src/roomsUtils.js'
 
 describe('filterByRoomType', function() {
   it('should be a function', function() {
@@ -51,20 +51,3 @@ describe('filterByRoomType', function() {
     expect(filter.length).to.deep.equal(0)
   })
 });
-describe('calculateBookingCost', function() {
-  it('should be a function', function() {
-    expect(calculateBookingCost).to.be.a('function')
-  });
-  it('should return the booking cost for one scheduled room', () => {
-    let room1Cost = 358.4
-    let totalCost = calculateBookingCost(roomsData,[room1Cost])
-    let expected = 358.4
-    expect(totalCost).to.deep.equal(expected)
-  })
-  it('should return the booking cost for multiple scheduled rooms', () => {
-    let room1Cost = 358.4
-    let room2Cost = 429.44
-    let totalCost = calculateBookingCost(roomsData,[room1Cost, room2Cost])
-    expect(totalCost).to.equal(room1Cost + room2Cost)
-  })
-})
